\name{BayDelC-package}
\alias{BayDelC-package}
\alias{BayDelC}
\docType{package}
\title{BayDelC: Bayesian Proxy System Models for Bottom-Water Oxygen}
\description{
Forward and inverse proxy-system models for the relationship between
bottom-water oxygen and benthic foraminiferal carbon-isotope offsets.
}
\details{
Use \code{forward_psm()} to simulate proxy values from oxygen and
\code{inverse_psm()} to reconstruct oxygen from proxy observations.
}
